+++ js/filters.js
@@ -1,9 +1,9 @@
-import {renderGallery} from './make-miniatures.js';
 import {photos} from './api.js';
+import {debouncedRenderGallery,sortRandomly,comparePhotos} from './util.js';
 
 const PICTURE_COUNT = 10;
 const Filter = {
-  DEFOULT: 'filter-default',
+  DEFAULT: 'filter-default',
   RANDOM: 'filter-random',
   DISCUSSED:'filter-discussed'
 };
@@ -13,37 +13,7 @@ const randomPhotosButton = document.querySelector('#filter-random');
 const defoultPhotosButton = document.querySelector('#filter-default');
 const filtersForm = document.querySelector('.img-filters__form');
 
-let currentFilter = Filter.DEFOULT;
-
-
-function debounce (callback, timeoutDelay = 500) {
-  let timeoutId;
-  return (...rest) => {
-    clearTimeout(timeoutId);
-    timeoutId = setTimeout(() => callback.apply(this, rest), timeoutDelay);
-  };
-}
-
-const debouncedRenderGallery = debounce(renderGallery);
-
-function comparePhotos (photo1, photo2) {
-  const rank1 = photo1.comments.length;
-  const rank2 = photo2.comments.length;
-
-  return rank2 - rank1;
-}
-
-function sortRandomly(arr){
-  let j, temp;
-  for(let i = arr.length - 1; i > 0; i--){
-    j = Math.floor(Math.random() * (i + 1));
-    temp = arr[j];
-    arr[j] = arr[i];
-    arr[i] = temp;
-  }
-  return arr;
-}
-
+let currentFilter = Filter.DEFAULT;
 
 filtersForm.addEventListener('click', (evt) => {
   if (!evt.target.classList.contains('img-filters__button')) {
